Split `MultiTablePaging` module and add haddocks
Splitting the state encoding functions to a separate module makes it
impossible to use them by mistake with the wrong bytestring (like the
Cassandra paging token).
