+++ changelog.d/4-docs/multitable-docs
@@ -1,0 +1,1 @@
+Add documentation of the multi-table paging abstraction
